requirements.txt
click==8.1.7
joblib==1.3.2
mysql-connector-python==8.1.0
nltk==3.8.1
Pillow==10.0.0
protobuf==4.21.12
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyqt5-tools==5.15.9.3.3
PyQt6-Qt6==6.5.2
python-dotenv==1.0.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
regex==2023.8.8
reportlab==4.0.4
tqdm==4.66.1
